Fixed up SimMain and changed StopEvent to take float instead of long. The program runs and seem to work according to spec

+++ src/deds/events/StopEvent.java
@@ -17,7 +17,7 @@ public class StopEvent extends Event
 	 *  
 	 * @param time when this event should occur
 	 */
-	public StopEvent( long time) {
+	public StopEvent( float time) {
 		super("Stop", time);
 	}
